New codes

+++ tempCodeRunnerFile.cpp
@@ -2,15 +2,15 @@
 using namespace std;
 int main()
 {
-	long long int t, n;
+	int t, n;
 	cin >> t;
 	while (t--)
 	{
 		cin >> n;
-		for (long long int i = 1; i <= (n / 2); i++)
+		for (int i = 1; i <= (n / 2); i++)
 		{
 			// cout << "bc" << i << endl;
-			long long int temp = i * i + 1;
+			int temp = i * i + 1;
 			if (temp / 2 > n)
 			{
 				cout << i - 1 << endl;
